refactor: Add trips_helper file that will contain the shared logic between different API's
